Fix  PlayReload compatibility problems
sbt.inc.Analysys -> sbt.internal.inc.Analysys

There are still some API migration TODOs here.
